models/open_model_8/requirements.txt
numerapi===2.19.1
python-dotenv===1.0.1
pandas===2.2.3
lightgbm===4.5.0
pyarrow===17.0.0
psutil===6.0.0
xgboost===2.1.3
catboost===1.2.7